# [Proposal] log: Record.Body as any